fix: stop deletion of EKS clusters when AWS controller terminated
The creation of EKS clusters includes an asyncronous function that
deletes AWS EKS resources if the controller is killed while creating
resources.  A provision to stop that process when creation is complete
was missing so, post create, if the controller was killed, an existing,
healthy cluster was being deleted.  This is fixed.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>