fix: remove warnings when installing threeport
The annotations error was due to the annotaions on the API server
service being at the config root rather than nested under metadata.
The TLS error was due to using Secrets of type: kubernetes.io/tls with
an empty tls.key.  This has been switched to an Opaque secret with only
the tls.crt for the CA on the controllers and agent.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>