refactor: move Threeport config code to pkg/cli/v0 package
The package `pkg/config/v0` contains the configuration abstractions for
CLI objects to make interacting with those objects more human-friendly.
The package `pkg/cli/v0` contains tptctl functionality.
The Threeport config lives on a user's filesystem and is used by tptctl
to connect to the Threeport API. It is far more sensible for this to
live in the cli package.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>