refactor: return config objects from all config abstraction methods
The config abstraction methods previously returned API objects which the
client (cmd/tptctl/cmd package) had to then translate into user-relevant
info.  This also did not give a user a way to get configs they could
modify and then submit as a part of a change.  This issue is remedied
with this commit.
This commit adds Get and Replace methods for each config
abstraction.
This commit removes the Describe methods.  They are no longer
needed.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>