refactor: remove describe cmd, return config objects from get cmd
This commit overhauls the tptctl CLI so that config objects from the
config package are returned to the user from the `get` command, rather
than select fields from the API objects.  This way, the user can
retrieve objects using tptctl and make modifications to them in order
to make changes since the input required is a config object.

This significantly cleans up the contract between tptctl and config
package.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>