refactor: use Config object for config package method receivers
Now each method for get, create, replace and delete objects in the
config package uses the Config object as the receiver.  Previously, it
had the Values object as the receiver while returning the Config
package.  This change was made for improved consistency and greater ease
of use by clients of this package.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>