fix: repair custom route association with handler funcs
Recently, an `IdParam` field was added to the `CustomRoute` object that
caused REST paths with an ID parameter to be routed to the same handler
function as those without an ID param.  This was a mistake since handler
funcs that accept ID params are always different functions.  This
mistake is corrected in this commit.

The SDK code generation for this file is also brought up-to-date with
the addition of the `ApiObjects` field in the `CustomRoute` struct.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>