feat: improve error message handling from threeport API gateways
Until now, the response processing in the client library has been
tailored to the response format from the Threeport API.  When there is
an intermediate proxy, such as an API gateway, if the error message was
in the response body, is was lost.  Now, the client library will
surface those errors in the response body to the user.

Signed-off-by: Rich Lander <lander2k2@protonmail.com>